Update remove_dir_check(): real option

+++ src/chrisbase/io.py
@@ -418,11 +418,11 @@ def remove_dir(path, ignore_errors=False) -> Path:
     return path
 
 
-def remove_dir_check(path, verbose=False, ignore_errors=True, file=None):
+def remove_dir_check(path, real=True, verbose=False, ignore_errors=True, file=None):
     path = Path(path)
     if verbose:
         print(f"- {str(path):<40}: {OX(path.exists())}", end='', file=file)
-    if path.exists():
+    if path.exists() and real:
         shutil.rmtree(path, ignore_errors=ignore_errors)
     if verbose:
         print(f" -> {OX(path.exists())}", file=file)
